move sortActionParams function to top-level action utils
